+++ apps/film-api/CHANGELOG.md
@@ -1,4 +1,11 @@
-# v0.1 - XX/01/2013
+# v0.1 - 13/01/2013
 - info handler
 - healthcheck handler
+
+# v0.2 - 10/02/2013
+- comment service connection
+- go 1.19 version
+
+# v0.3 - 13/02/2013
+- client side streaming connection